Finding cities with temperature higher than, not rainy cities and scheduler

+++ src/main/java/com/weather/imgw/Weather_App/api/weather/WeatherStationDto.java
@@ -9,24 +9,13 @@ public class WeatherStationDto {
     private long id_stacji;
     private String stacja;
     private LocalDate data_pomiaru;
-    private double godzina_pomiaru;
+    private int godzina_pomiaru;
     private double temperatura;
     private double predkosc_wiatru;
-    private double kierunek_wiatru;
+    private int kierunek_wiatru;
     private double wilgotnosc_wzgledna;
     private double suma_opadu;
     private double cisnienie;
 
-    //    {
-//        id_stacji: "12295",
-//                stacja: "Białystok",
-//            data_pomiaru: "2020-03-08",
-//            godzina_pomiaru: "9",
-//            temperatura: "5.1",
-//            predkosc_wiatru: "2",
-//            kierunek_wiatru: "250",
-//            wilgotnosc_wzgledna: "97.2",
-//            suma_opadu: "0000000002.9",
-//            cisnienie: "1017.6"
-//    },
+
 }
